+++ source/Internal/Framework/Smarty/Configuration/SmartySettingsDataProvider.php
@@ -9,12 +9,15 @@ declare(strict_types=1);
 
 namespace OxidEsales\EshopCommunity\Internal\Framework\Smarty\Configuration;
 
-use OxidEsales\EshopCommunity\Core\Registry;
+use OxidEsales\EshopCommunity\Internal\Framework\Smarty\Extension\SmartyTemplateHandlerInterface;
 use OxidEsales\EshopCommunity\Internal\Framework\Smarty\SmartyContextInterface;
 
 class SmartySettingsDataProvider implements SmartySettingsDataProviderInterface
 {
-    public function __construct(private SmartyContextInterface $context)
+    public function __construct(
+        private SmartyContextInterface $context,
+        private SmartyTemplateHandlerInterface $smartyTemplateHandler
+    )
     {
     }
 
@@ -34,11 +37,11 @@ class SmartySettingsDataProvider implements SmartySettingsDataProviderInterface
             'cache_dir' => $compilePath,
             'template_dir' => $this->context->getTemplateDirectories(),
             'compile_id' => $this->getTemplateCompileId(),
-            'default_template_handler_func' => [Registry::getUtilsView(), '_smartyDefaultTemplateHandler'],
+            'default_template_handler_func' => [$this->smartyTemplateHandler, 'handleTemplate'],
             'debugging' => $this->context->getTemplateEngineDebugMode(),
             'compile_check' => $this->context->getTemplateCompileCheckMode(),
-            'php_handling' => $this->context->getTemplatePhpHandlingMode(),
-            'security' => false,
+            'php_handling' => (int) $this->context->getTemplatePhpHandlingMode(),
+            'security' => false
         ];
     }