+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
     - `convertToFormattedDbDate()`
     - `convertToPseudoHtml()`
   - `OxidEsales\EshopCommunity\Core\UtilsView:parseThroughSmarty()`
+  - `OxidEsales\EshopCommunity\Core\UtilsView:_smartyDefaultTemplateHandler()`
 
 ### Fixed
 - Partly revert `OxidEsales\EshopCommunity\Core\Autoload\ModuleAutoload`